+++ bark_ml/observers/nearest_observer.hpp
@@ -46,7 +46,8 @@ using bark::models::dynamic::StateDefinition::X_POSITION;
 using bark::models::dynamic::StateDefinition::Y_POSITION;
 using bark::models::dynamic::StateDefinition::THETA_POSITION;
 using bark::models::dynamic::StateDefinition::VEL_POSITION;
-using ObservedState = Eigen::Matrix<double, 1, Eigen::Dynamic>;
+// needs to be float to be compatible with tf.flaot32 and np.float32
+using ObservedState = Eigen::Matrix<float, 1, Eigen::Dynamic>;
 using bark::commons::transformation::FrenetPosition;
 using State = Eigen::Matrix<double, Eigen::Dynamic, 1>;
